# Perf hash nodes