--- a/validity-prover/src/trees/merkle_tree/sql_node_hash.rs
+++ b/validity-prover/src/trees/merkle_tree/sql_node_hash.rs
@@ -167,3 +167,369 @@ impl<V: Leafable + Serialize + DeserializeOwned> SqlNodeHashes<V> {
         Ok(())
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use rand::Rng;
+    use sqlx::postgres::PgPoolOptions;
+
+    use super::{BitPath, SqlNodeHashes};
+    use crate::trees::utils::bit_path::BitPath as BitPathOrig;
+    use intmax2_zkp::utils::leafable::Leafable;
+
+    type TestValue = u32;
+
+    fn setup_test() -> String {
+        dotenv::dotenv().ok();
+        std::env::var("DATABASE_URL").unwrap()
+    }
+
+    // helper func
+    fn collect_paths(path: &BitPath) -> Vec<BitPath> {
+        let mut paths = vec![];
+        let mut current_path = *path;
+        while !current_path.is_empty() {
+            paths.push(current_path);
+            current_path.pop();
+        }
+        paths
+    }
+
+    #[tokio::test]
+    async fn test_prove_empty_tree() -> anyhow::Result<()> {
+        let database_url = setup_test();
+        let pool = PgPoolOptions::new()
+            .max_connections(5)
+            .connect(&database_url)
+            .await?;
+
+        let mut rng = rand::thread_rng();
+        let tag = rng.gen();
+        let height = 10;
+        let node_hashes = SqlNodeHashes::<TestValue>::new(pool.clone(), tag, height);
+
+        let timestamp = 1000;
+        let mut tx = pool.begin().await?;
+
+        for index in [0, 1, 5, 10, 100] {
+            let proof = node_hashes.prove(&mut tx, timestamp, index).await?;
+
+            let expected_length = height;
+            assert_eq!(
+                proof.siblings.len(),
+                expected_length,
+                "Proof for index {} should have {} siblings",
+                index,
+                expected_length
+            );
+
+            let mut path = BitPath::new(height as u32, index);
+            path.reverse();
+            let paths = collect_paths(&path);
+
+            for (i, sibling) in proof.siblings.iter().enumerate() {
+                let sibling_path = paths[i].sibling();
+                let expected_zero_hash = node_hashes.zero_hashes[sibling_path.len() as usize];
+                assert_eq!(
+                    *sibling, expected_zero_hash,
+                    "Sibling hash at position {} for index {} should be zero hash",
+                    i, index
+                );
+            }
+        }
+
+        tx.rollback().await?;
+        Ok(())
+    }
+
+    #[tokio::test]
+    async fn test_prove_sequential_entries() -> anyhow::Result<()> {
+        let database_url = setup_test();
+        let pool = PgPoolOptions::new()
+            .max_connections(5)
+            .connect(&database_url)
+            .await?;
+
+        let mut rng = rand::thread_rng();
+        let tag = rng.gen();
+        let height = 10;
+        let node_hashes = SqlNodeHashes::<TestValue>::new(pool.clone(), tag, height);
+
+        let timestamp = 1000;
+        let mut tx = pool.begin().await?;
+
+        let num_entries = 16;
+        let mut expected_hashes = Vec::new();
+
+        for i in 0..num_entries {
+            let bit_path = BitPathOrig::new(height as u32, i);
+            let hash_value = TestValue::empty_leaf().hash();
+            expected_hashes.push((bit_path, hash_value)); // store expected values
+            node_hashes
+                .save_node(&mut tx, timestamp, bit_path, hash_value)
+                .await?;
+        }
+
+        for index in 0..num_entries {
+            let proof = node_hashes.prove(&mut tx, timestamp, index).await?;
+
+            let mut path = BitPath::new(height as u32, index);
+            path.reverse();
+            let paths = collect_paths(&path);
+
+            assert_eq!(
+                proof.siblings.len(),
+                height,
+                "Proof for index {} should have {} siblings",
+                index,
+                height
+            );
+
+            for (i, sibling) in proof.siblings.iter().enumerate() {
+                let sibling_path = paths[i].sibling();
+
+                let sibling_index = sibling_path.value();
+                if sibling_index >= num_entries && i > height - 4 {
+                    let expected_zero_hash = node_hashes.zero_hashes[sibling_path.len() as usize];
+                    assert_eq!(
+                        *sibling, expected_zero_hash,
+                        "Sibling hash at position {} for index {} should be zero hash",
+                        i, index
+                    );
+                }
+            }
+        }
+
+        tx.rollback().await?;
+        Ok(())
+    }
+
+    #[tokio::test]
+    async fn test_prove_sparse_entries() -> anyhow::Result<()> {
+        let database_url = setup_test();
+        let pool = PgPoolOptions::new()
+            .max_connections(5)
+            .connect(&database_url)
+            .await?;
+
+        let mut rng = rand::thread_rng();
+        let tag = rng.gen();
+        let height = 10;
+        let node_hashes = SqlNodeHashes::<TestValue>::new(pool.clone(), tag, height);
+
+        let timestamp = 1000;
+        let mut tx = pool.begin().await?;
+
+        let sparse_indices = [100, 200, 500, 1000];
+        let hash_value = TestValue::empty_leaf().hash();
+
+        for &i in &sparse_indices {
+            let bit_path = BitPathOrig::new(height as u32, i);
+            node_hashes
+                .save_node(&mut tx, timestamp, bit_path, hash_value)
+                .await?;
+        }
+
+        for &index in &sparse_indices {
+            let proof = node_hashes.prove(&mut tx, timestamp, index).await?;
+            let expected_length = height;
+            assert_eq!(
+                proof.siblings.len(),
+                expected_length,
+                "Proof for sparse index {} should have {} siblings",
+                index,
+                expected_length
+            );
+        }
+
+        tx.rollback().await?;
+        Ok(())
+    }
+
+    #[tokio::test]
+    async fn test_prove_boundary_indices() -> anyhow::Result<()> {
+        let database_url = setup_test();
+        let pool = PgPoolOptions::new()
+            .max_connections(5)
+            .connect(&database_url)
+            .await?;
+
+        let mut rng = rand::thread_rng();
+        let tag = rng.gen();
+        let height = 10;
+        let node_hashes = SqlNodeHashes::<TestValue>::new(pool.clone(), tag, height);
+
+        let timestamp = 1000;
+        let mut tx = pool.begin().await?;
+
+        let boundary_indices = [0, (1 << height) - 1, (1 << (height - 1))];
+        let hash_value = TestValue::empty_leaf().hash();
+
+        for &i in &boundary_indices {
+            let bit_path = BitPathOrig::new(height as u32, i);
+            node_hashes
+                .save_node(&mut tx, timestamp, bit_path, hash_value)
+                .await?;
+        }
+
+        for &index in &boundary_indices {
+            let proof = node_hashes.prove(&mut tx, timestamp, index).await?;
+
+            let expected_length = height;
+            assert_eq!(
+                proof.siblings.len(),
+                expected_length,
+                "Proof for boundary index {} should have {} siblings",
+                index,
+                expected_length
+            );
+        }
+
+        tx.rollback().await?;
+        Ok(())
+    }
+
+    #[tokio::test]
+    async fn test_prove_timestamps() -> anyhow::Result<()> {
+        let database_url = setup_test();
+        let pool = PgPoolOptions::new()
+            .max_connections(5)
+            .connect(&database_url)
+            .await?;
+
+        let mut rng = rand::thread_rng();
+        let tag = rng.gen();
+        let height = 10;
+        let node_hashes = SqlNodeHashes::<TestValue>::new(pool.clone(), tag, height);
+
+        let mut tx = pool.begin().await?;
+
+        let timestamps = [500, 800, 1200];
+        let test_index = 42;
+        let hash_value = TestValue::empty_leaf().hash();
+
+        for &ts in &timestamps {
+            let bit_path = BitPathOrig::new(height as u32, test_index);
+            node_hashes
+                .save_node(&mut tx, ts, bit_path, hash_value)
+                .await?;
+        }
+
+        for ts in timestamps {
+            let proof = node_hashes.prove(&mut tx, ts, test_index).await?;
+            let expected_length = height;
+            assert_eq!(
+                proof.siblings.len(),
+                expected_length,
+                "Proof at timestamp {} should have {} siblings",
+                ts,
+                expected_length
+            );
+        }
+
+        tx.rollback().await?;
+        Ok(())
+    }
+
+    #[tokio::test]
+    async fn test_prove_all_zero_hashes() -> anyhow::Result<()> {
+        let database_url = setup_test();
+        let pool = PgPoolOptions::new()
+            .max_connections(5)
+            .connect(&database_url)
+            .await?;
+
+        let mut rng = rand::thread_rng();
+        let tag = rng.gen();
+        let height = 10;
+        let node_hashes = SqlNodeHashes::<TestValue>::new(pool.clone(), tag, height);
+
+        let timestamp = 1000;
+        let mut tx = pool.begin().await?;
+
+        let large_indices = [
+            (1 << height) - 2,         // near maximum
+            (1 << (height - 1)) + 123, // intermediate
+            (1 << height) / 3,         // other value
+        ];
+
+        for &index in &large_indices {
+            let proof = node_hashes.prove(&mut tx, timestamp, index).await?;
+
+            let mut path = BitPath::new(height as u32, index);
+            path.reverse();
+            let paths = collect_paths(&path);
+
+            for (i, sibling) in proof.siblings.iter().enumerate() {
+                let sibling_path = paths[i].sibling();
+                let expected_zero_hash = node_hashes.zero_hashes[sibling_path.len() as usize];
+                assert_eq!(
+                    *sibling, expected_zero_hash,
+                    "Sibling hash at position {} for index {} should be zero hash",
+                    i, index
+                );
+            }
+        }
+
+        tx.rollback().await?;
+        Ok(())
+    }
+
+    #[tokio::test]
+    async fn test_prove_mixed_zero_and_real_hashes() -> anyhow::Result<()> {
+        let database_url = setup_test();
+        let pool = PgPoolOptions::new()
+            .max_connections(5)
+            .connect(&database_url)
+            .await?;
+
+        let mut rng = rand::thread_rng();
+        let tag = rng.gen();
+        let height = 10;
+        let node_hashes = SqlNodeHashes::<TestValue>::new(pool.clone(), tag, height);
+
+        let timestamp = 1000;
+        let mut tx = pool.begin().await?;
+
+        let test_index = 42;
+        let mut path = BitPath::new(height as u32, test_index);
+        path.reverse();
+
+        let paths = collect_paths(&path);
+
+        for i in (0..paths.len()).step_by(2) {
+            let path_to_save = paths[i];
+            let sibling_path = path_to_save.sibling();
+
+            let custom_hash = TestValue::empty_leaf().hash();
+            node_hashes
+                .save_node(&mut tx, timestamp, sibling_path, custom_hash)
+                .await?;
+        }
+
+        let proof = node_hashes.prove(&mut tx, timestamp, test_index).await?;
+
+        for (i, sibling) in proof.siblings.iter().enumerate() {
+            let sibling_path = paths[i].sibling();
+
+            if i % 2 == 0 {
+                let saved_hash = TestValue::empty_leaf().hash();
+                assert_eq!(
+                    *sibling, saved_hash,
+                    "Sibling hash at position {} should be custom hash",
+                    i
+                );
+            } else {
+                let expected_zero_hash = node_hashes.zero_hashes[sibling_path.len() as usize];
+                assert_eq!(
+                    *sibling, expected_zero_hash,
+                    "Sibling hash at position {} should be zero hash",
+                    i
+                );
+            }
+        }
+
+        tx.rollback().await?;
+        Ok(())
+    }
+}