## Change Summary
This PR enhances performance optimizations for query hash_nodes and adds comprehensive test coverage:

* Added `bulk_get_sibling_hashes` method to retrieve multiple node hashes with a single SQL query
* Improved the prove method to use bulk retrieval instead of individual hash lookups, reducing database access
* Added `test_speed_incremental_merkle_tree_prove` test to measure `prove` performance
* Added unit tests

## Benchmark Result Summary
performance has significantly improved. The implementation of bulk hash retrieval reduced the time for 5000 proof generations from 63.1 seconds to 7.9 seconds (an 87% reduction).

## Benchmark Results

benchmark script: `validity-prover/src/trees/merkle_tree/mod.rs:test_speed_incremental_merkle_tree_prove`

### Before
```
running 1 test
test trees::merkle_tree::tests::test_speed_incremental_merkle_tree_prove has been running for over 60 seconds
SqlIncrementMerkleTree.prove: 5000 times, 32 height, 63.127032334 seconds
test trees::merkle_tree::tests::test_speed_incremental_merkle_tree_prove ... ok

test result: ok. 1 passed; 0 failed; 0 ignored; 0 measured; 17 filtered out; finished in 63.15s
```

```
 schema_name |             table_name              |                index_name                | scans  | index_size | total_size | index_ratio | tuples_read | tuples_fetched
-------------+-------------------------------------+------------------------------------------+--------+------------+------------+-------------+-------------+----------------
 public      | hash_nodes                          | idx_hash_nodes_lookup                    | 159990 | 2592 kB    | 4536 kB    |       57.14 |           0 |              0
 public      | hash_nodes                          | hash_nodes_pkey                          |     21 | 1624 kB    | 4536 kB    |       35.80 |          11 |              0
 public      | leaves_len                          | leaves_len_pkey                          |      1 | 48 kB      | 4536 kB    |        1.06 |           1 |              1
 public      | leaves                              | leaves_pkey                              |      1 | 56 kB      | 4536 kB    |        1.23 |           1 |              1
 public      | leaves_len                          | idx_leaves_len_lookup                    |      1 | 72 kB      | 4536 kB    |        1.59 |           0 |              0
 public      | _sqlx_migrations                    | _sqlx_migrations_pkey                    |      0 | 16 kB      | 4536 kB    |        0.35 |           0 |              0
 public      | tx_tree_roots                       | tx_tree_roots_pkey                       |      0 | 8192 bytes | 4536 kB    |        0.18 |           0 |              0
 public      | validity_proofs                     | validity_proofs_pkey                     |      0 | 8192 bytes | 4536 kB    |        0.18 |           0 |              0
 public      | prover_tasks                        | prover_tasks_pkey                        |      0 | 8192 bytes | 4536 kB    |        0.18 |           0 |              0
 public      | validity_state                      | validity_state_pkey                      |      0 | 8192 bytes | 4536 kB    |        0.18 |           0 |              0
 public      | deposit_leaf_events                 | deposit_leaf_events_pkey                 |      0 | 8192 bytes | 4536 kB    |        0.18 |           0 |              0
 public      | full_blocks                         | full_blocks_pkey                         |      0 | 8192 bytes | 4536 kB    |        0.18 |           0 |              0
 public      | deposit_leaf_events                 | idx_deposit_leaf_events_deposit_hash     |      0 | 8192 bytes | 4536 kB    |        0.18 |           0 |              0
 public      | deposit_leaf_events                 | idx_deposit_leaf_events_block_tx         |      0 | 8192 bytes | 4536 kB    |        0.18 |           0 |              0
 public      | full_blocks                         | idx_full_blocks_block_tx                 |      0 | 8192 bytes | 4536 kB    |        0.18 |           0 |              0
 public      | observer_deposit_sync_eth_block_num | observer_deposit_sync_eth_block_num_pkey |      0 | 8192 bytes | 4536 kB    |        0.18 |           0 |              0
 public      | observer_block_sync_eth_block_num   | observer_block_sync_eth_block_num_pkey   |      0 | 8192 bytes | 4536 kB    |        0.18 |           0 |              0
 public      | indexed_leaves                      | indexed_leaves_pkey                      |      0 | 8192 bytes | 4536 kB    |        0.18 |           0 |              0
 public      | indexed_leaves                      | idx_indexed_leaves_lookup                |      0 | 8192 bytes | 4536 kB    |        0.18 |           0 |              0
 public      | indexed_leaves                      | idx_indexed_leaves_timestamp             |      0 | 8192 bytes | 4536 kB    |        0.18 |           0 |              0
 public      | prover_tasks                        | idx_prover_tasks_assigned_status         |      0 | 8192 bytes | 4536 kB    |        0.18 |           0 |              0
 public      | tx_tree_roots                       | idx_tx_tree_roots_block_number           |      0 | 8192 bytes | 4536 kB    |        0.18 |           0 |              0
(22 rows)
```


### After

```
running 1 test
SqlIncrementMerkleTree.prove: 5000 times, 32 height, 7.857474416 seconds
test trees::merkle_tree::tests::test_speed_incremental_merkle_tree_prove ... ok

test result: ok. 1 passed; 0 failed; 0 ignored; 0 measured; 17 filtered out; finished in 7.88s
```

```
 schema_name |             table_name              |                index_name                | scans  | index_size | total_size | index_ratio | tuples_read | tuples_fetched
-------------+-------------------------------------+------------------------------------------+--------+------------+------------+-------------+-------------+----------------
 public      | hash_nodes                          | idx_hash_nodes_lookup                    | 156224 | 2592 kB    | 4536 kB    |       57.14 |           0 |              0
 public      | hash_nodes                          | hash_nodes_pkey                          |   4148 | 1624 kB    | 4536 kB    |       35.80 |        4058 |              0
 public      | leaves_len                          | leaves_len_pkey                          |      1 | 48 kB      | 4536 kB    |        1.06 |           1 |              1
 public      | leaves                              | leaves_pkey                              |      1 | 56 kB      | 4536 kB    |        1.23 |           1 |              1
 public      | leaves_len                          | idx_leaves_len_lookup                    |      1 | 72 kB      | 4536 kB    |        1.59 |           0 |              0
 public      | _sqlx_migrations                    | _sqlx_migrations_pkey                    |      0 | 16 kB      | 4536 kB    |        0.35 |           0 |              0
 public      | tx_tree_roots                       | tx_tree_roots_pkey                       |      0 | 8192 bytes | 4536 kB    |        0.18 |           0 |              0
 public      | validity_proofs                     | validity_proofs_pkey                     |      0 | 8192 bytes | 4536 kB    |        0.18 |           0 |              0
 public      | prover_tasks                        | prover_tasks_pkey                        |      0 | 8192 bytes | 4536 kB    |        0.18 |           0 |              0
 public      | validity_state                      | validity_state_pkey                      |      0 | 8192 bytes | 4536 kB    |        0.18 |           0 |              0
 public      | deposit_leaf_events                 | deposit_leaf_events_pkey                 |      0 | 8192 bytes | 4536 kB    |        0.18 |           0 |              0
 public      | full_blocks                         | full_blocks_pkey                         |      0 | 8192 bytes | 4536 kB    |        0.18 |           0 |              0
 public      | deposit_leaf_events                 | idx_deposit_leaf_events_deposit_hash     |      0 | 8192 bytes | 4536 kB    |        0.18 |           0 |              0
 public      | deposit_leaf_events                 | idx_deposit_leaf_events_block_tx         |      0 | 8192 bytes | 4536 kB    |        0.18 |           0 |              0
 public      | full_blocks                         | idx_full_blocks_block_tx                 |      0 | 8192 bytes | 4536 kB    |        0.18 |           0 |              0
 public      | observer_deposit_sync_eth_block_num | observer_deposit_sync_eth_block_num_pkey |      0 | 8192 bytes | 4536 kB    |        0.18 |           0 |              0
 public      | observer_block_sync_eth_block_num   | observer_block_sync_eth_block_num_pkey   |      0 | 8192 bytes | 4536 kB    |        0.18 |           0 |              0
 public      | indexed_leaves                      | indexed_leaves_pkey                      |      0 | 8192 bytes | 4536 kB    |        0.18 |           0 |              0
 public      | indexed_leaves                      | idx_indexed_leaves_lookup                |      0 | 8192 bytes | 4536 kB    |        0.18 |           0 |              0
 public      | indexed_leaves                      | idx_indexed_leaves_timestamp             |      0 | 8192 bytes | 4536 kB    |        0.18 |           0 |              0
 public      | prover_tasks                        | idx_prover_tasks_assigned_status         |      0 | 8192 bytes | 4536 kB    |        0.18 |           0 |              0
 public      | tx_tree_roots                       | idx_tx_tree_roots_block_number           |      0 | 8192 bytes | 4536 kB    |        0.18 |           0 |              0
(22 rows)
```